requirements.txt
paramiko
flask-cors
flask
python-dotenv
google-generativeai


#pip install -r requirements.txt